[Core/Spell] Chi wave exploit fix
An addition to Jens' fix, since he forgot about the actual spell that does the visual, which used to.

Current behaviour:

Chi wave cannot be casted directly on a PvP flagged player unless you're PvP flagged as well, because the spell tells you that the target is PvP enabled.
If you cast Chi wave on yourself, and then it chooses a PvP flagged player as the bounce hit, it will flag the monk.
If the monk is PvP flagged while casting Chi wave on a PvP disabled player and then it bounces to a PvP flagged player or creature, the primary target (the initial player, not the monk) won't be flagged.